feat(entities): added contacts selector component in forms prodile and create peroject
