+++ src/entities/form-create-project/ui/form-create-project.module.scss
@@ -107,20 +107,6 @@
 	}
 }
 
-.contacts {
-	width: 100%;
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-
-	&_selects {
-		display: grid;
-		gap: 8px;
-		grid-template-columns: 228px auto;
-		height: 70px;
-	}
-}
-
 .buttons {
 	margin-top: 20px;
 	display: flex;
@@ -130,75 +116,6 @@
 	}
 }
 
-.fields {
-	display: flex;
-	flex-direction: column;
-	padding: 32px;
-	gap: 16px;
-	box-sizing: border-box;
-	// min-width: 752px;
-	border-radius: 12px;
-	border: 1px solid $--border-default;
-	margin-bottom: 24px;
-
-	@media screen and (max-width: 779px) {
-		border: none;
-	}
-
-	&__addContactWrapper {
-		width: 100%;
-		display: flex;
-		flex-direction: row;
-		gap: 8px;
-		@media screen and (max-width: 779px) {
-			flex-direction: column;
-		}
-	}
-	&__addContactType {
-		width: 228px;
-		height: 40px;
-		border-radius: 8px;
-		border: 1px solid $--border-default;
-		background: $--bg-primary;
-		padding: 8px 12px;
-		color: $--Foreground-default;
-		font-feature-settings:
-			'liga' off,
-			'clig' off;
-		font-family: $--open-sans-font;
-		font-size: 16px;
-		font-style: normal;
-		font-weight: 400;
-		line-height: 24px; /* 150% */
-		letter-spacing: 0.25px;
-		@media screen and (max-width: 779px) {
-			width: 100%;
-		}
-	}
-
-	&__addContactTypeWrapper {
-		display: flex;
-		flex-direction: column;
-	}
-
-	&__addContactTextValue {
-		min-height: 40px;
-		width: 100%;
-	}
-
-	&__addContactTypeLabel {
-		color: $--Foreground-subtitle;
-		font-feature-settings:
-			'liga' off,
-			'clig' off;
-		font-family: $--open-sans-font;
-		font-size: 14px;
-		font-style: normal;
-		font-weight: 400;
-		line-height: 19px; /* 135.714% */
-		letter-spacing: 0.25px;
-	}
-}
 .checkboxError {
 	padding: 0;
 	margin: 0;
