+++ src/entities/form-profile-edit/ui/form-profile-edit.module.scss
@@ -49,7 +49,7 @@
 		font-family: $--open-sans-font;
 		font-size: 16px;
 		font-style: normal;
-		font-weight: 400;
+		font-weight: 700;
 		line-height: 24px; /* 150% */
 		letter-spacing: 0.25px;
 	}
