+++ src/shared/utils/index.ts
@@ -1,2 +1,3 @@
 export { getColorTag } from './get-tag-color/get-tag-color';
 export { getDate } from './get-current-date/get-current-date';
+export { mergeContacts } from './merge-contacts/merge-contacts';
