+++ src/widgets/contacts-selector/index.ts
@@ -1,0 +1,1 @@
+export { ContactsSelector } from './ui/contacts-selector';
