+++ src/widgets/contacts-selector/ui/contacts-selector.module.scss
@@ -1,0 +1,125 @@
+@import '@/shared/assets/fonts/fonts.css';
+@import '@/shared/style/variables';
+
+.input_list {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 8px;
+
+	&_title {
+		margin: 0;
+		font-family: $--open-sans-font;
+		font-weight: 700;
+		font-size: 16px;
+		font-style: normal;
+		line-height: 150%;
+	}
+    &_titleProfile {
+        font-weight: 400;
+    }
+}
+.contacts {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 8px;
+
+	&_selects {
+		display: grid;
+		gap: 8px;
+		grid-template-columns: 228px auto;
+		height: 70px;
+	}
+}
+
+.fields {
+	display: flex;
+	flex-direction: column;
+	padding: 32px;
+	gap: 16px;
+	box-sizing: border-box;
+	// min-width: 752px;
+	border-radius: 12px;
+	border: 1px solid $--border-default;
+	margin-bottom: 24px;
+
+	@media screen and (max-width: 779px) {
+		border: none;
+	}
+
+	&__addContactWrapper {
+		width: 100%;
+		display: flex;
+		flex-direction: row;
+		gap: 8px;
+		@media screen and (max-width: 779px) {
+			flex-direction: column;
+		}
+	}
+	&__addContactType {
+		width: 228px;
+		height: 40px;
+		border-radius: 8px;
+		border: 1px solid $--border-default;
+		background: $--bg-primary;
+		padding: 8px 12px;
+		color: $--Foreground-default;
+		font-feature-settings:
+			'liga' off,
+			'clig' off;
+		font-family: $--open-sans-font;
+		font-size: 16px;
+		font-style: normal;
+		font-weight: 400;
+		line-height: 24px; /* 150% */
+		letter-spacing: 0.25px;
+		@media screen and (max-width: 779px) {
+			width: 100%;
+		}
+	}
+
+	&__addContactTypeWrapper {
+		display: flex;
+		flex-direction: column;
+	}
+
+	&__addContactTextValue {
+		min-height: 40px;
+		width: 100%;
+	}
+
+	&__addContactTypeLabel {
+		color: $--Foreground-subtitle;
+		font-feature-settings:
+			'liga' off,
+			'clig' off;
+		font-family: $--open-sans-font;
+		font-size: 14px;
+		font-style: normal;
+		font-weight: 400;
+		line-height: 19px; /* 135.714% */
+		letter-spacing: 0.25px;
+	}
+}
+
+.specialists {
+	width: 100%;
+	font-family: $--open-sans-font;
+	font-weight: 400;
+	display: flex;
+	flex-direction: column;
+	gap: 24px;
+
+	&_toggle {
+		display: flex;
+		gap: 16px;
+		align-items: center;
+	}
+
+	&_buttons {
+		display: flex;
+		justify-content: space-between;
+		align-items: center;
+	}
+}
